# Implement constraints in pyro.param statements