… tutorial/source/gmm.ipynb