Closes #749 

This adds new syntax
```py
sigma = pyro.param('sigma', torch.tensor(1.0),  # <--- no need to requires_grad=True
                   constraint=constraints.positive)

assert pyro.param('sigma') > 0  # <--- no need to specify constraint
```
Note that constraints are not yet compatible with param store load/save operations. This is blocked by https://github.com/probtorch/pytorch/issues/138, after which load/save should work fine.

## Tested

- [x] Convert integration tests to use `constraints.positive`
- [x] Convert some tests to use `constraints.unit_interval`
- [x] Measure performance overhead

## Benchmarks

| | before (sec) | after (sec) |
| --- | --- | ---
| test_exponential_gamma | 107.52 | 107.43 |
| NormalNormalTests | 32.06 | 31.54 | 

Pair coded with @eb8680.