@fritzo 2018-03-20
OK, ready to merge. I added some benchmarking of some tests and found no speed penalty.